[ADD] 캐릭터 Select

+++ Server/Server/RoomManager.h
@@ -10,10 +10,10 @@ public:
 	static RoomManager& Instance();
 
 private:
-	map<uint32, GameRoomRef> _rooms;
-	uint32 _curId = 0;
+	map<uint8, GameRoomRef> _rooms;
+	uint8 _curId = 0;
 	
 public:
 	void Init();
-	GameRoom* Find(uint32& id);
+	GameRoom* Find(uint8& id);
 };
